# perf: improve http cache performance