--- a/Cargo.toml
+++ b/Cargo.toml
@@ -149,6 +149,7 @@ convert_case = "0.6.0"
 rand = "0.8.5"
 tailcall-macros = { path = "tailcall-macros" }
 tailcall-tracker = { path = "tailcall-tracker", optional = true }
+tailcall-http-cache = {path = "tailcall-http-cache"}
 tonic-types = "0.11.0"
 datatest-stable = "0.2.9"
 tokio-test = "0.4.4"
@@ -211,7 +212,9 @@ members = [
     "tailcall-query-plan",
     "tailcall-fixtures",
     "tailcall-upstream-grpc",
-    "tailcall-tracker"]
+    "tailcall-tracker",
+    "tailcall-http-cache"
+]
 
 # Boost execution_spec snapshot diffing performance
 [profile.dev.package]