--- a/src/core/runtime.rs
+++ b/src/core/runtime.rs
@@ -42,10 +42,11 @@ pub mod test {
     use std::time::Duration;
 
     use anyhow::{anyhow, Result};
-    use http_cache_reqwest::{Cache, CacheMode, HttpCache, HttpCacheOptions, MokaManager};
+    use http_cache_reqwest::{Cache, CacheMode, HttpCache, HttpCacheOptions};
     use hyper::body::Bytes;
     use reqwest::Client;
     use reqwest_middleware::{ClientBuilder, ClientWithMiddleware};
+    use tailcall_http_cache::HttpCacheManager;
     use tokio::io::{AsyncReadExt, AsyncWriteExt};
 
     use crate::cli::javascript;
@@ -97,7 +98,7 @@ pub mod test {
             if upstream.http_cache {
                 client = client.with(Cache(HttpCache {
                     mode: CacheMode::Default,
-                    manager: MokaManager::default(),
+                    manager: HttpCacheManager::default(),
                     options: HttpCacheOptions::default(),
                 }))
             }