--- a/tailcall-http-cache/Cargo.toml
+++ b/tailcall-http-cache/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+name = "tailcall-http-cache"
+version = "0.1.0"
+edition = "2021"
+publish = false
+
+[dependencies]
+http-cache-reqwest = { version = "0.13.0", default-features = false, features = [] }
+http-cache-semantics = { version = "1.0.2", default-features = false, features = ["with_serde", "reqwest"]}
+moka = { version = "0.12.7", default-features = false, features = ["future"] }
+serde = "1.0.200"
+async-trait = "0.1.80"
+
+[dev-dependencies]
+tokio = {version = "1.37.0", features = ["full"]}
+url = { version = "2.5.0", features = ["serde"] }
+reqwest = { workspace = true }
+http = "0.2.12"
+http-cache = "0.18.0"
+anyhow = { workspace = true }