--- a/tailcall-http-cache/src/lib.rs
+++ b/tailcall-http-cache/src/lib.rs
@@ -1,0 +1,3 @@
+mod http_cache_manager;
+
+pub use http_cache_manager::HttpCacheManager;