**Summary:**  
- created separate workspace, tailcall-http-cache which contains the optimised cache manager.
- bench results:
 before:   [38.054 µs 38.204 µs 38.428 µs]
 after:      [18.419 µs 18.432 µs 18.449 µs]
 performance improvement : 51%.

**Issue Reference(s):**  
Related #1852

**Build & Testing:**

- [ ] I ran `cargo test` successfully.
- [ ] I have run `./lint.sh --mode=fix` to fix all linting issues raised by `./lint.sh --mode=check`.

**Checklist:**

- [ ] I have added relevant unit & integration tests.
- [ ] I have updated the [documentation] accordingly.
- [ ] I have performed a self-review of my code.
- [ ] PR follows the naming convention of `<type>(<optional scope>): <title>`

[documentation]: https://github.com/tailcallhq/tailcallhq.github.io/tree/develop/docs